cpp01

+++ ex01/zombieHorde.cpp
@@ -4,10 +4,9 @@ Zombie* zombieHorde( int N, std::string name ) {
 
 	Zombie *zombies = new Zombie[N];
 
-	for (size_t i = 0; i < N; i++) {
+	for (int i = 0; i < N; i++) {
 		zombies[i].setName(name);
 	}
-
 
 	return zombies;
 }
